backend/requirements.txt
Django==4
# djangorestframework==3
flwr==1
# requests==2.28

# Django>=4
djangorestframework>=3
# flwr>=1
requests>=2.28
